// sdk/ts/loopxComponents.ts
// Code generated by goctl. DO NOT EDIT.

export interface SocialInfo {
	social_type: string
	uuid: string
}

export interface UserInfo {
	id: number
	username: string
	email: string
	avatar?: string
	walletAddress: string
	primaryWalletAddress: string
	state: number
	following_count: number
	follower_count: number
	social: Array<SocialInfo>
}

export interface ReplyStatus {
	code: number
	msg: string
}

export interface UserLoginReq {
}
export interface UserLoginReqParams {
	login_type: string
	wallet_address: string
	email: string
	nonce: string
	social_type: string
	social_id: string
}
export interface UserLoginReqHeaders {
	wallet_sign: string
}

export interface UserInfoReply {
	id: number
	username: string
	email: string
	avatar?: string
	walletAddress: string
	primaryWalletAddress: string
	state: number
	following_count: number
	follower_count: number
	social: Array<SocialInfo>
	code: number
	msg: number
	access_token?: string
}

export interface CaptchaReq {
	walletAddress: string
	ts: number
}

export interface CaptchaReply {
	code: number
	msg: string
	nonce: string
}

export interface AccountReq {
}
export interface AccountReqParams {
}

export interface UserSettingReq {
	username?: string
	walletAddress?: string
	email?: string
	avatar?: string
	primaryWalletAddress?: string
}

export interface NotificationReply {
	code: number
	msg: string
}

export interface UserFollowingReq {
}
export interface UserFollowingReqParams {
}

export interface UserFollowingReply {
	code: number
	msg: string
	users: Array<UserInfo>
}

export interface UserFollowerReq {
}
export interface UserFollowerReqParams {
}

export interface UserFollowerReply {
	code: number
	msg: string
	users: Array<UserInfo>
}

export interface LikeReq {
}
export interface LikeReqParams {
}

export interface LikeReply {
	code: number
	msg: string
	address_list: Array<string>
}
export interface LikeReplyParams {
}
export interface LikeReplyHeaders {
}

export interface MetricContractReply {
	code: number
	msg: string
}

export interface MetricTokenReply {
	code: number
	msg: string
}

export interface FeedsListReq {
	period?: number
}

export interface FeedsListReply {
	code: number
	msg: string
}

export interface CommentInfo {
	id: number
	topic_id: number
	topic_type: number
	reply_id: number
	reply_type: number
	from_user: string
	content: string
	like_count: number
	date: number
}

export interface CommentListReq {
}
export interface CommentListReqParams {
}

export interface CommentDeleteReq {
}
export interface CommentDeleteReqParams {
}

export interface CommentListReply {
	code: number
	msg: string
	comments: Array<CommentInfo>
}

export interface SearchInfo {
	address: string
	score: number
}

export interface SearchReq {
}
export interface SearchReqParams {
	q: string
	type: string
}

export interface SearchReply {
	code: number
	msg: string
	search_info: Array<SearchInfo>
}